+++ gruntfile.js
@@ -1,15 +1,33 @@
 module.exports = function(grunt) {
 	"use strict";
 
-	// load what we need
-	grunt.loadNpmTasks('grunt-contrib-uglify');
+  var fs = require("fs");
+  var packageFile = grunt.file.readJSON("package.json");
+
+	grunt.loadNpmTasks("grunt-contrib-uglify");
+  grunt.loadNpmTasks("grunt-template");
+
 
 	var config = {
+    template: {
+      generate: {
+        options: {
+          data: {
+            __VERSION__: packageFile.version,
+            __DATA__: fs.readFileSync("source/data.js", "utf8")
+          }
+        },
+        files: {
+          "dist/crs.js": ["source/source-crs.js"],
+          "dist/jquery.crs.js": ["source/source-jquery.crs.js"]
+        }
+      }
+    },
 		uglify: {
 			standalone: {
 				files: {
-					'source/crs.min.js': 'source/crs.js',
-					'source/jquery.crs.min.js': 'source/jquery.crs.js'
+					"dist/crs.min.js": "dist/crs.js",
+					"dist/jquery.crs.min.js": "dist/jquery.crs.js"
 				},
 				options: {
 					report: "min",
@@ -20,6 +38,6 @@ module.exports = function(grunt) {
 	};
 
 	grunt.initConfig(config);
-	grunt.registerTask('default', ['uglify']);
-	grunt.registerTask('generate', ['uglify']);
+	grunt.registerTask("default", ["template:generate", "uglify"]);
+	grunt.registerTask("generate", ["template:generate", "uglify"]);
 };
